feat: domain, view 폴더로 구조 정리
